Refactor scenes directory
