+++ src/game/scenes/MainScene.js
@@ -1,23 +1,23 @@
 import Phaser from 'phaser';
-import backgroundImage from '../assets/images/background.jpg';
-import authorImage from '../assets/images/author2.png';
-import logoImage from '../assets/images/logo.png';
-import buttonImage from '../assets/images/button.png';
-import buttonClickedImage from '../assets/images/button-clicked.png';
-import buttonHoverImage from '../assets/images/button-hover.png';
-import lifeImage from '../assets/images/life.png';
-import shotSprite from '../assets/spritesheets/shot2.png';
-import playerDieSprite from '../assets/spritesheets/player-die.png';
-import playerMoveSprite from '../assets/spritesheets/player-move.png';
-import playerShotSprite from '../assets/spritesheets/player-shot.png';
-import zombie1DieSprite from '../assets/spritesheets/zombie1-die.png';
-import zombie1MoveSprite from '../assets/spritesheets/zombie1-move.png';
-import zombie2DieSprite from '../assets/spritesheets/zombie2-die.png';
-import zombie2MoveSprite from '../assets/spritesheets/zombie2-move.png';
-import pixelFontImage from '../assets/font/font.png';
-import pixelFontXML from '../assets/font/font.xml';
+import backgroundImage from '../../assets/images/background.jpg';
+import authorImage from '../../assets/images/author2.png';
+import logoImage from '../../assets/images/logo.png';
+import buttonImage from '../../assets/images/button.png';
+import buttonClickedImage from '../../assets/images/button-clicked.png';
+import buttonHoverImage from '../../assets/images/button-hover.png';
+import lifeImage from '../../assets/images/life.png';
+import shotSprite from '../../assets/spritesheets/shot2.png';
+import playerDieSprite from '../../assets/spritesheets/player-die.png';
+import playerMoveSprite from '../../assets/spritesheets/player-move.png';
+import playerShotSprite from '../../assets/spritesheets/player-shot.png';
+import zombie1DieSprite from '../../assets/spritesheets/zombie1-die.png';
+import zombie1MoveSprite from '../../assets/spritesheets/zombie1-move.png';
+import zombie2DieSprite from '../../assets/spritesheets/zombie2-die.png';
+import zombie2MoveSprite from '../../assets/spritesheets/zombie2-move.png';
+import pixelFontImage from '../../assets/font/font.png';
+import pixelFontXML from '../../assets/font/font.xml';
 
-import createAnimations from '../game/helpers/animations-creator';
+import createAnimations from '../helpers/animations-creator';
 
 class MainScene extends Phaser.Scene {
   constructor() {
